% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authorize.R
\name{authorize}
\alias{authorize}
\title{Log in to Banner}
\usage{
authorize()
}
\value{
An \code{rvest} session
}
\description{
\code{authorize()} uses values stored during setup to authorize a browsing session in Banner.
}
\examples{
\dontrun{
  my_session <- authorize()
}
}
