% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bannr-package.R
\docType{package}
\name{bannr-package}
\alias{bannr}
\alias{bannr-package}
\title{bannr: Pull Data from Banner}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Log in to Banner Web to pull data on course enrollments and attendances. Using rvest and tidyverse tools, bannr helps course instructors get clean, actionable data on their courses, their students, and other information that can otherwise only be navigated via the web.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://jmclawson.github.io/bannr/}
  \item \url{https://github.com/jmclawson/bannr}
  \item Report bugs at \url{https://github.com/jmclawson/bannr/issues}
}

}
\author{
\strong{Maintainer}: James Clawson \email{clawson@gmail.com} (\href{https://orcid.org/0009-0008-6756-0607}{ORCID})

}
\keyword{internal}
