% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten_roster.R
\name{flatten_rosters}
\alias{flatten_rosters}
\title{Convert list of course data to a table}
\usage{
flatten_rosters(rosters)
}
\arguments{
\item{rosters}{A list of course data}
}
\value{
A table of course data. The current \code{rvest} session is attached as an attribute, used for chaining functions together.
}
\description{
Convert list of course data to a table
}
\examples{
\dontrun{
  my_classes <- authorize() |>
    get_rosters() |>
    get_attendances() |>
    flatten_rosters() |>
    process_attendances()
}
}
