% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rosters.R
\name{get_rosters}
\alias{get_rosters}
\title{Get roster data}
\usage{
get_rosters(courses, .session = NULL, term = NULL)
}
\arguments{
\item{courses}{A course table}

\item{.session}{(optional) An active \code{rvest} session. If this value is undefined, the session will be derived from an attribute set on \code{courses}}

\item{term}{(optional) The term to be used in this query. If undefined, this value will be inherited from \code{rosters} or from the environmental variable defined at setup.}
}
\value{
A list of items, with one item per course. Each item will identify the course CRN, catalog number, title, schedule, and a table showing the roster of student enrollments. The current \code{rvest} session is attached as an attribute, used for chaining functions together.
}
\description{
\code{get_rosters()} uses an active \code{rvest} session to find data on each course, including a table of students enrolled.
}
\examples{
\dontrun{
  my_classes <- authorize() |>
    get_rosters()
}
}
