% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_webpage.R
\name{make_attendance_dashboard}
\alias{make_attendance_dashboard}
\title{Create webpage for attendance taking}
\usage{
make_attendance_dashboard(df, filename = "attendance-dashboard.html")
}
\arguments{
\item{df}{A table of course data, such as the one derived after \code{\link[=flatten_rosters]{flatten_rosters()}}. This table needs to have columns that include \code{crn}, \code{catalog}, \code{course}, \code{days}, \code{start}, and \code{end}.}

\item{filename}{The path and file name for saving the HTML file.}
}
\description{
\code{make_attendance_dashboard()} makes a clean webpage to speed the process of taking attendance at the start of class. While Banner requires loading five webpages to change tabs, choose the link for attendance, and select the term, course, and date, \code{bannr}'s attendance dashboard links directly to the final step. Links to course rosters are also provided.
}
\examples{
\dontrun{
  authorize() |> 
    get_rosters() |>
    flatten_rosters() |>
    make_attendance_dashboard("take-attendance.html")
}
}
