% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_attendances.R
\name{process_attendances}
\alias{process_attendances}
\title{Add summaries of attendance}
\usage{
process_attendances(attendance_rosters, include_fridays = TRUE)
}
\arguments{
\item{attendance_rosters}{A table of course enrollments and attendances}

\item{include_fridays}{Whether to count Fridays among a course's viable days of attendance. This parameter is useful when teaching an online or hybrid class when Fridays serve as marker of that week's attendance, but where Friday is otherwise missing from course records as the day of meeting.}
}
\value{
A table of course enrollments and attendances, with added columns showing each student's current rate of attendance and the last date attended. The current \code{rvest} session is attached as an attribute, used for chaining functions together.
}
\description{
Add summaries of attendance
}
\examples{
\dontrun{
  my_classes <- authorize() |>
    get_rosters() |>
    get_attendances() |>
    flatten_rosters() |>
    process_attendances()
}
}
