% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authorize.R
\name{setup}
\alias{setup}
\title{Set required values}
\usage{
setup(reset = FALSE)
}
\arguments{
\item{reset}{Indicates whether previous values should be ignored.}
}
\value{
A message indicating completion
}
\description{
\code{setup()} stores key values for logging into Banner, including the institutional URL, the current term, and others. The user will also be prompted to store a login username and password in the system keychain.
}
\section{Required values}{

Values requested at this stage include the following:
\enumerate{
\item The \strong{base url} of a Banner instance, used when navigating by web browser
\item The \strong{current term} used internally by Banner. This might take the form of something like "202420" for the second semester of the 2023-2024 academic year.
\item For the \strong{login page}, any url details added to the base url. Once the base url has been entered, this value has a suggested value.
\item For the \emph{course rosters page}, any url details added to the base url. Once the base url has been entered, this value has a suggested value.
\item For the \emph{attendance rosters page}, any url details added to the base url. Once the base url has been entered, this value has a suggested value.
\item The \strong{username} and \strong{password} used to log in to the system.
}
}

\section{Security considerations}{

The first five values, above, should not be considered sensitive information. They are stored as environmental variables. The \strong{username} and \strong{password} requested in the final step are given heightened security consideration and are saved in your computer system's keyring, provided it has been set up.
}

