% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_course_table.R
\name{get_course_table}
\alias{get_course_table}
\title{Get a table of courses}
\usage{
get_course_table(.session)
}
\arguments{
\item{.session}{An active \code{rvest} session}
}
\value{
A tibble showing the CRN, title, subject, catalog number, and enrollment for each course. The current \code{rvest} session is attached as an attribute, used for chaining functions together.
}
\description{
Get a table of courses
}
\examples{
\dontrun{
  my_courses <- authorize() |>
    get_course_table()
}
}
