% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authorize.R
\name{setup}
\alias{setup}
\title{Set required values}
\usage{
setup(reset = FALSE)
}
\arguments{
\item{reset}{Indicates whether previous values should be ignored.}
}
\value{
A message indicating completion
}
\description{
\code{setup()} stores key values for logging into Banner, including the institutional URL, the current term, and others. The user will also be prompted to store a login username and password in the system keychain.
}
